Added correct calculation for nutrition in AddIntakeCommandHandler and some refactoring.

+++ NutritionTracker/NutritionTracker.Application/Commands/AddIntake/AddIntakeCommand.cs
@@ -12,6 +12,6 @@ namespace NutritionTracker.Application.Commands.AddIntake
 		public DateOnly DateOfIntake {  get; set; }
 		public Guid UserId { get; set; }
 		public int FoodId { get; set; }
-		public float FoodAmount { get; set; }
+		public double FoodAmount { get; set; }
 	}
 }
